Make pagination more robust
Previously the pagination would enable its buttons based on the number
of words. So if you save/confirm one word you have less than page-size
words on the page and hence the next button would be disabled. Which
isn't really very smart.

Now the prev and next buttons are enabled/disabled in the
fetch-words-success event handler. If we managed to fetch page-size
words then we assume that there are going to be more words. If we
received less that page-size words we assume that there aren't any
more words and we disable the next button.

Fixes #27