Enable search with ^ and $ "wildcards"
This makes the search more predictable for the user, just type a word
and it works sort of like google, specify ^ or $ to limit the search.

However the way this is implemented using a LIKE and a search term
that starts with a % wildcard, this really kills performance at least
in the case of global words. Need to find a solition for this.

Fixes #48