Improve pagination
The offset is now the real offset into the words instead of just a
multiplicator for the page-size. When some words are saved/removed
from a page then the offset is increased by the number of visible
words so that the next page starts at the word that hasn't been seen
before, i.e. we will not get any overlaps.

Fixes #39