Add a simple migration for the local words
Create a new table based on a create table statement without select
into, so that the indexes are also created properly.

On updrade rename the original table and aggregate the different grade
rows into one row in the new table.

On downgrade drop the new table and just rename the old table to the
original. This of course assumes that the old table wasn't removed.