Add some rudimentary tests for the db
These tests depend on an existing db and some existing data in the db.
In other words they are _very_ brittle and work with the current MySQL
test db. Needless to say that they fail miserably in CI where the test
db is not available.

I tried to port the test db to h2 and got pretty far: I migrated the
schema and convinced the connection pool to connect to the in-memory
db. But in the end it appears that some of the queries are not
compatible with h2, e.g. SELECT (CASE column WHEN "foo" THEN "bar" END
or ON DUPLICATE KEY UPDATE. So this seems pretty futile.

With SQLite compatibility is apparently even worse.